SLIM-2077-added-test-with-mocked-soap-client
