+++ osgp/platform/osgp-secret-management/src/integration-test/resources/test-requests/getSecrets.xml
@@ -2,5 +2,6 @@
   <ns:DeviceId>E0054002019112319</ns:DeviceId>
   <ns:SecretTypes>
     <ns:SecretType>E_METER_AUTHENTICATION_KEY</ns:SecretType>
+    <ns:SecretType>E_METER_ENCRYPTION_KEY_UNICAST</ns:SecretType>
   </ns:SecretTypes>
 </ns:getSecretsRequest>
