+++ osgp/platform/osgp-secret-management/src/integration-test/resources/test-responses/getSecrets.xml
@@ -1,1 +1,13 @@
-<test>TODO</test>
+<ns2:getSecretsResponse xmlns:ns2="http://www.opensmartgridplatform.org/schemas/security/secretmanagement/2020/05">
+    <ns2:Result>OK</ns2:Result>
+    <ns2:TypedSecrets>
+        <ns2:TypedSecret>
+            <ns2:Type>E_METER_AUTHENTICATION_KEY</ns2:Type>
+            <ns2:Secret>74efc062231e81c9e006bb56c5dec38631210c5073511606a203ba748fcdc794</ns2:Secret>
+        </ns2:TypedSecret>
+        <ns2:TypedSecret>
+            <ns2:Type>E_METER_ENCRYPTION_KEY_UNICAST</ns2:Type>
+            <ns2:Secret>3dca51832c70e372460796ca01acbab769fd330c9b936246a01d4e97f8c5bc26</ns2:Secret>
+        </ns2:TypedSecret>
+    </ns2:TypedSecrets>
+</ns2:getSecretsResponse>
