+++ osgp/platform/osgp-secret-management/src/main/java/org/opensmartgridplatform/secretmgmt/application/Main.java
@@ -11,5 +11,4 @@ public class Main extends SpringBootServletInitializer {
     public static void main(final String[] args) {
         SpringApplication.run(Main.class, args);
     }
-
 }
